DRY in TemplateType implementations
