+++ src/Type/Generic/TemplateObjectType.php
@@ -3,29 +3,18 @@
 namespace PHPStan\Type\Generic;
 
 use PHPStan\TrinaryLogic;
-use PHPStan\Type\CompoundType;
-use PHPStan\Type\IntersectionType;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\ObjectWithoutClassType;
 use PHPStan\Type\Traits\UndecidedComparisonCompoundTypeTrait;
 use PHPStan\Type\Type;
-use PHPStan\Type\UnionType;
-use PHPStan\Type\VerbosityLevel;
 
 final class TemplateObjectType extends ObjectType implements TemplateType
 {
 
 	use UndecidedComparisonCompoundTypeTrait;
-
-	private TemplateTypeScope $scope;
-
-	private string $name;
-
-	private TemplateTypeStrategy $strategy;
-
-	private ObjectType $bound;
-
-	private TemplateTypeVariance $variance;
+	use TemplateTypeTrait {
+		isSubTypeOf as isSubTypeOfTrait;
+	}
 
 	public function __construct(
 		TemplateTypeScope $scope,
@@ -44,43 +33,13 @@ final class TemplateObjectType extends ObjectType implements TemplateType
 		$this->bound = new ObjectType($class);
 	}
 
-	public function getName(): string
+	public function isSubTypeOf(Type $type): TrinaryLogic
 	{
-		return $this->name;
-	}
+		if ($type instanceof ObjectWithoutClassType) {
+			return TrinaryLogic::createYes();
+		}
 
-	public function getScope(): TemplateTypeScope
-	{
-		return $this->scope;
-	}
-
-	public function getBound(): Type
-	{
-		return $this->bound;
-	}
-
-	public function describe(VerbosityLevel $level): string
-	{
-		$basicDescription = function () use ($level): string {
-			return sprintf(
-				'%s of %s',
-				$this->name,
-				parent::describe($level)
-			);
-		};
-
-		return $level->handle(
-			$basicDescription,
-			$basicDescription,
-			function () use ($basicDescription): string {
-				return sprintf('%s (%s, %s)', $basicDescription(), $this->scope->describe(), $this->isArgument() ? 'argument' : 'parameter');
-			}
-		);
-	}
-
-	public function isArgument(): bool
-	{
-		return $this->strategy->isArgument();
+		return $this->isSubTypeOfTrait($type);
 	}
 
 	public function toArgument(): TemplateType
@@ -92,114 +51,6 @@ final class TemplateObjectType extends ObjectType implements TemplateType
 			$this->name,
 			$this->getClassName()
 		);
-	}
-
-	public function isValidVariance(Type $a, Type $b): TrinaryLogic
-	{
-		return $this->variance->isValidVariance($a, $b);
-	}
-
-	public function subtract(Type $type): Type
-	{
-		return $this;
-	}
-
-	public function getTypeWithoutSubtractedType(): Type
-	{
-		return $this;
-	}
-
-	public function changeSubtractedType(?Type $subtractedType): Type
-	{
-		return $this;
-	}
-
-	public function equals(Type $type): bool
-	{
-		return $type instanceof self
-			&& $type->scope->equals($this->scope)
-			&& $type->name === $this->name
-			&& parent::equals($type);
-	}
-
-	public function isAcceptedBy(Type $acceptingType, bool $strictTypes): TrinaryLogic
-	{
-		return $this->isSubTypeOf($acceptingType);
-	}
-
-	public function accepts(Type $type, bool $strictTypes): TrinaryLogic
-	{
-		return $this->strategy->accepts($this, $type, $strictTypes);
-	}
-
-	public function isSuperTypeOf(Type $type): TrinaryLogic
-	{
-		if ($type instanceof CompoundType) {
-			return $type->isSubTypeOf($this);
-		}
-
-		return $this->getBound()->isSuperTypeOf($type)
-			->and(TrinaryLogic::createMaybe());
-	}
-
-	public function isSubTypeOf(Type $type): TrinaryLogic
-	{
-		if ($type instanceof UnionType || $type instanceof IntersectionType) {
-			return $type->isSuperTypeOf($this);
-		}
-
-		if ($type instanceof ObjectWithoutClassType) {
-			return TrinaryLogic::createYes();
-		}
-
-		if (!$type instanceof TemplateType) {
-			return $type->isSuperTypeOf($this->getBound());
-		}
-
-		if ($this->equals($type)) {
-			return TrinaryLogic::createYes();
-		}
-
-		if ($type->getBound()->isSuperTypeOf($this->getBound())->no() &&
-			$this->getBound()->isSuperTypeOf($type->getBound())->no()) {
-			return TrinaryLogic::createNo();
-		}
-
-		return TrinaryLogic::createMaybe();
-	}
-
-	public function inferTemplateTypes(Type $receivedType): TemplateTypeMap
-	{
-		if ($receivedType instanceof UnionType || $receivedType instanceof IntersectionType) {
-			return $receivedType->inferTemplateTypesOn($this);
-		}
-
-		if (
-			$receivedType instanceof TemplateType
-			&& $this->getBound()->isSuperTypeOf($receivedType->getBound())->yes()
-		) {
-			return new TemplateTypeMap([
-				$this->name => $receivedType,
-			]);
-		}
-
-		if ($this->getBound()->isSuperTypeOf($receivedType)->yes()) {
-			return new TemplateTypeMap([
-				$this->name => TemplateTypeHelper::generalizeType($receivedType),
-			]);
-		}
-
-		return TemplateTypeMap::createEmpty();
-	}
-
-	public function getReferencedTemplateTypes(TemplateTypeVariance $positionVariance): array
-	{
-		return [new TemplateTypeReference($this, $positionVariance)];
-	}
-
-	public function getVariance(): TemplateTypeVariance
-	{
-		return $this->variance;
 	}
 
 	/**
